Set up route for user profile and add file upload in HTML
Added inputs into `lib/view/profile.html` to provide front-end
capabilities of uploading files for an artist to be able to upload
cover photo, profile photo, and audio in `.png` and `.mp3` file
formats, respectively.

For image upload, put a `form` in a `div` and set that `div`'s
background image. For audio, set the `input` to `hidden` and created
a `label`.

Also set up route for the user (i.e non-artist)